Refactor cat.cs to use multiple files using #:include directive
