+++ global.json
@@ -1,6 +1,6 @@
 {
   "sdk": {
-    "version": "10.0.100",
+    "version": "10.0.300-preview",
     "rollForward": "latestFeature",
     "allowPrerelease": true
   }
